# specs/001-agent-orchestration/contracts/api.yaml
openapi: 3.0.3
info:
  title: Multi-Agent Code Development Orchestration Platform API
  description: REST API for orchestrating AI agents in code development workflows
  version: 1.0.0
  contact:
    name: API Support
    email: support@agent-orchestration.local
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.agent-orchestration.com/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /agents:
    get:
      summary: List registered agents
      description: Returns a list of all registered AI agents with their status
      operationId: listAgents
      tags:
        - Agents
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
                    minimum: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Register a new agent
      description: Register a new AI agent with the platform
      operationId: registerAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistration'
      responses:
        '201':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid agent configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agentId}:
    get:
      summary: Get agent details
      description: Returns detailed information about a specific agent
      operationId: getAgent
      tags:
        - Agents
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the agent
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update agent configuration
      description: Update the configuration of an existing agent
      operationId: updateAgent
      tags:
        - Agents
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid agent configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Unregister agent
      description: Remove an agent from the platform
      operationId: unregisterAgent
      tags:
        - Agents
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the agent
      responses:
        '204':
          description: Agent unregistered successfully
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Agent:
      type: object
      required:
        - id
        - name
        - provider
        - model
        - capabilities
        - status
      properties:
        id:
          type: string
          description: Unique identifier for the agent
        name:
          type: string
          description: Human-readable name
        provider:
          type: string
          enum: [anthropic, openai, github]
          description: AI provider name
        model:
          type: string
          description: Model identifier
        capabilities:
          type: array
          items:
            type: string
          description: List of supported capabilities
        status:
          type: string
          enum: [active, inactive, error]
          description: Current operational status
        config:
          type: object
          description: Provider-specific configuration
        lastHealthCheck:
          type: string
          format: date-time
          description: Timestamp of last health check
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentRegistration:
      type: object
      required:
        - id
        - provider
        - model
        - config
      properties:
        id:
          type: string
          description: Unique identifier for the agent
        name:
          type: string
          description: Human-readable name
        provider:
          type: string
          enum: [anthropic, openai, github]
        model:
          type: string
        config:
          type: object
          description: Provider-specific configuration including API keys
        capabilities:
          type: array
          items:
            type: string
          description: List of supported capabilities
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
        capabilities:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT